This pull request introduces several improvements and refinements to the
client application, focusing on user experience, backend integration,
and configuration updates. The most significant changes include enhanced
error handling and validation for game creation, improved employee data
loading via API, and updated configuration for production deployment.
Additionally, UI adjustments and minor code cleanups have been made for
clarity and usability.

**Game Creation Flow Improvements**
* Enhanced error handling and retry logic in `CreateGameScreen`: The
game creation form now validates input (e.g., rounds must be a number
within a valid range), displays error messages, and allows users to
retry if creation fails. The loading dialog is properly disposed of, and
the form's control flags are reset after errors.
[[1]](diffhunk://#diff-ab9fc22dc9e8b473ee59d3c925b0dff1614924d7026087f9a2963e00e3b58d14L9-L87)
[[2]](diffhunk://#diff-ab9fc22dc9e8b473ee59d3c925b0dff1614924d7026087f9a2963e00e3b58d14L102-R194)
[[3]](diffhunk://#diff-ab9fc22dc9e8b473ee59d3c925b0dff1614924d7026087f9a2963e00e3b58d14R212)
[[4]](diffhunk://#diff-ab9fc22dc9e8b473ee59d3c925b0dff1614924d7026087f9a2963e00e3b58d14L252-R324)
[[5]](diffhunk://#diff-ab9fc22dc9e8b473ee59d3c925b0dff1614924d7026087f9a2963e00e3b58d14L262-R354)
[[6]](diffhunk://#diff-ab9fc22dc9e8b473ee59d3c925b0dff1614924d7026087f9a2963e00e3b58d14L206-R269)
* The rounds range in the form label is updated to "5 - 20" for clearer
instructions.

**Employee Data Integration**
* Employee data for the current game is now loaded asynchronously from
the backend API (`/employee/game/{gameId}`), replacing previous reliance
on local state. This improves accuracy and ensures up-to-date
information.
[[1]](diffhunk://#diff-36d32944f51c67cb0662ded13f7fe044b63e7b8dc6b82b4eab6556e776c7a99fL479-R482)
[[2]](diffhunk://#diff-36d32944f51c67cb0662ded13f7fe044b63e7b8dc6b82b4eab6556e776c7a99fR579-R617)
[[3]](diffhunk://#diff-36d32944f51c67cb0662ded13f7fe044b63e7b8dc6b82b4eab6556e776c7a99fL597-R640)

**Configuration and Deployment Updates**
* The API and WebSocket server domains/ports have been updated from
`localhost:7032` to `oak.tail025bf6.ts.net:443` in both
`Client/Program.cs` and `Client/appsettings.json`, preparing the client
for production deployment.
[[1]](diffhunk://#diff-2ca465883857100b1e083046e0a0e05dfa016daf2b95a1e0401265a7ae55c70bL21-R27)
[[2]](diffhunk://#diff-dd6f014642c3a7689853c9ad919fe8c0b3c0ac575ba0ed043fb25db629852d6bL4-R10)
* Added a publish profile for folder deployment in
`Client/Properties/PublishProfiles/FolderProfile.pubxml`.
* Added a `.devcontainer.json` for standardized development environments
with Nix and VS Code Copilot extensions.

**UI and Usability Enhancements**
* Improved color scheme in the main window for better readability and
aesthetics.
* Adjusted button layout and added an error message label to the game
creation form for clearer feedback.
[[1]](diffhunk://#diff-ab9fc22dc9e8b473ee59d3c925b0dff1614924d7026087f9a2963e00e3b58d14L252-R324)
[[2]](diffhunk://#diff-ab9fc22dc9e8b473ee59d3c925b0dff1614924d7026087f9a2963e00e3b58d14L262-R354)

**Miscellaneous**
* The changelog workflow now also runs on the `dev` branch.
* Minor type and enum renaming for clarity (e.g., `EnrollInFormation` to
`EnrollEmployee`, `Salary` type changed to `decimal`).
[[1]](diffhunk://#diff-623aa0b55de0a0dcb9d196da0b93bf8e673224c13448daba28b4a7fced431579L13-R14)
[[2]](diffhunk://#diff-623aa0b55de0a0dcb9d196da0b93bf8e673224c13448daba28b4a7fced431579L39-R40)
[[3]](diffhunk://#diff-36d32944f51c67cb0662ded13f7fe044b63e7b8dc6b82b4eab6556e776c7a99fL773-R819)
[[4]](diffhunk://#diff-36d32944f51c67cb0662ded13f7fe044b63e7b8dc6b82b4eab6556e776c7a99fL820-R859)

Let me know if you want a deeper explanation of any specific change!

+++ Server/Hubs/Records/GameOverview.cs
@@ -36,7 +36,7 @@ public record ConsultantOverview(
 public sealed record EmployeeOverview(
     int Id,
     string Name,
-    double Salary,
+    decimal Salary,
     ICollection<SkillOverview> Skills
 );
